requirements.txt
flask
opencv-python
numpy
dlib
Pillow